pipeline-service/requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
polars>=0.20.0
pandas>=2.2.0
ydata-profiling>=4.17.0
great-expectations>=0.18.0
rapidfuzz>=3.6.0
chardet>=5.2.0
dateparser>=1.2.0
recordlinkage>=0.15
pandera>=0.18.0
scikit-learn>=1.4.0
python-multipart>=0.0.9
supabase>=2.3.0
httpx>=0.27.0
pyarrow>=15.0.0
openpyxl>=3.1.0
python-dotenv>=1.0.0
numpy>=1.26.0
